[Bronze V] Title: 이상한 기호, Time: 112 ms, Memory: 14324 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.kt
@@ -1,6 +1,6 @@
 fun main() = with(System.`in`.bufferedReader()) {
     val arr = readLine().split(" ")
-    println(calculate(arr[0].toInt(), arr[1].toInt()))
+    println(calculate(arr[0].toLong(), arr[1].toLong()))
 }
 
-private fun calculate(a : Int, b : Int) : Int = (a + b) * (a - b)
+private fun calculate(a : Long, b : Long) : Long = (a + b) * (a - b)
